+++ examples/reminders/lib/base/common_blocs/coordinator_bloc.dart
@@ -11,9 +11,11 @@ import 'package:rx_bloc/rx_bloc.dart';
 import 'package:rx_bloc_list/models.dart';
 import 'package:rxdart/rxdart.dart';
 
+import '../models/counter/increment_operation.dart';
 import '../models/reminder/reminder_model.dart';
 
 part 'coordinator_bloc.rxb.g.dart';
+
 part 'coordinator_bloc_extensions.dart';
 
 // ignore: one_member_abstracts
@@ -24,7 +26,8 @@ abstract class CoordinatorEvents {
 
   void reminderCreated(Result<ReminderModel> reminderResult);
 
-  void reminderUpdated(Result<ReminderModel> reminderResult);
+  void reminderUpdated(
+      Result<IdentifiablePair<ReminderModel>> reminderPairResult);
 }
 
 abstract class CoordinatorStates {
@@ -35,7 +38,7 @@ abstract class CoordinatorStates {
   Stream<Result<ReminderModel>> get onReminderDeleted;
 
   @RxBlocIgnoreState()
-  Stream<Result<ReminderModel>> get onReminderUpdated;
+  Stream<Result<IdentifiablePair<ReminderModel>>> get onReminderUpdated;
 
   @RxBlocIgnoreState()
   Stream<Result<ReminderModel>> get onReminderCreated;
@@ -57,7 +60,8 @@ class CoordinatorBloc extends $CoordinatorBloc {
   Stream<Result<ReminderModel>> get onReminderDeleted => _$reminderDeletedEvent;
 
   @override
-  Stream<Result<ReminderModel>> get onReminderUpdated => _$reminderUpdatedEvent;
+  Stream<Result<IdentifiablePair<ReminderModel>>> get onReminderUpdated =>
+      _$reminderUpdatedEvent;
 }
 
 extension CoordinatingTasksX on CoordinatorBlocType {
@@ -74,23 +78,38 @@ extension CoordinatingTasksX on CoordinatorBlocType {
   ///
   /// Based on the result of the callback [onUpdateOperation] the updated task
   ///  will/will not be removed from value of the provided [subject].
-  Stream<ManagedList<ReminderModel>> mapReminderManageEventsWithLatestFrom(
+  Stream<ManagedListCounterOperation> mapReminderManageEventsWithLatestFrom(
     Stream<List<ReminderModel>> reminderList, {
-    required Future<ManageOperation> Function(ReminderModel model,ReminderModel? identifiableInList)
+    required Future<ManageOperation> Function(ReminderModel model)
         operationCallback,
   }) =>
       Rx.merge([
-        states.onReminderCreated.whereSuccess().withLatestFromIdentifiableList(
+        states.onReminderCreated
+            .whereSuccess()
+            .withLatestFromIdentifiableList(
               reminderList,
               operationCallback: operationCallback,
-            ),
-        states.onReminderDeleted.whereSuccess().withLatestFromIdentifiableList(
+            )
+            .map((managedList) => ManagedListCounterOperation<ReminderModel>(
+                managedList: managedList,
+                counterOperation: CounterOperation.create)),
+        states.onReminderDeleted
+            .whereSuccess()
+            .withLatestFromIdentifiableList(
               reminderList,
-              operationCallback: (reminder,identifiableInList) async => ManageOperation.remove,
-            ),
-        states.onReminderUpdated.whereSuccess().withLatestFromIdentifiableList(
+              operationCallback: (reminder) async => ManageOperation.remove,
+            )
+            .map((managedList) => ManagedListCounterOperation<ReminderModel>(
+                managedList: managedList,
+                counterOperation: CounterOperation.delete)),
+        states.onReminderUpdated
+            .whereSuccess()
+            .withLatestFromIdentifiablePairList(
               reminderList,
               operationCallback: operationCallback,
-            ),
+            )
+            .map((managedList) => ManagedListCounterOperation<ReminderModel>(
+                managedList: managedList,
+                counterOperation: CounterOperation.update)),
       ]);
 }