+++ examples/reminders/lib/base/common_ui_components/app_reminder_tile.dart
@@ -56,6 +56,7 @@ class _AppReminderTileState extends State<AppReminderTile> {
               .events
               .update(widget.reminder.copyWith(
                 title: _textEditingController.text,
+                // completeUpdated: false,
               ));
 
           widget.onTitleChanged?.call(_textEditingController.text);
@@ -109,8 +110,33 @@ class _AppReminderTileState extends State<AppReminderTile> {
       ));
 
   ActionPane _buildActionPane(BuildContext context) => ActionPane(
+        extentRatio: 0.7,
         motion: const ScrollMotion(),
         children: [
+          SlidableAction(
+            onPressed: (context) =>
+                context.read<ReminderManageBlocType>().events.update(
+                      widget.reminder.copyWith(
+                        complete: !widget.reminder.complete,
+                        // completeUpdated: true,
+                      ),
+                    ),
+            backgroundColor: widget.reminder.complete
+                ? context.designSystem.colors.actionButtonCompleteColor
+                : context.designSystem.colors.actionButtonInCompleteColor,
+            foregroundColor: context.designSystem.colors.canvasColor,
+            icon: Icons.check_box_outlined,
+            label: widget.reminder.complete
+                ? context.l10n.incomplete
+                : context.l10n.complete,
+          ),
+          SlidableAction(
+            onPressed: null,
+            backgroundColor: context.designSystem.colors.activeButtonColor,
+            foregroundColor: context.designSystem.colors.canvasColor,
+            icon: Icons.edit,
+            label: context.l10n.edit,
+          ),
           SlidableAction(
             onPressed: (context) => context
                 .read<ReminderManageBlocType>()
@@ -121,13 +147,6 @@ class _AppReminderTileState extends State<AppReminderTile> {
             foregroundColor: context.designSystem.colors.canvasColor,
             icon: Icons.delete,
             label: context.l10n.delete,
-          ),
-          SlidableAction(
-            onPressed: null,
-            backgroundColor: context.designSystem.colors.activeButtonColor,
-            foregroundColor: context.designSystem.colors.canvasColor,
-            icon: Icons.edit,
-            label: context.l10n.edit,
           ),
         ],
       );
@@ -150,6 +169,7 @@ class _AppReminderTileState extends State<AppReminderTile> {
           .events
           .update(widget.reminder.copyWith(
             dueDate: date,
+            // completeUpdated: false,
           ));
 
       widget.onDueDateChanged?.call(date);