+++ examples/reminders/lib/feature_reminder_manage/blocs/reminder_manage_bloc.rxb.g.dart
@@ -42,8 +42,8 @@ abstract class $ReminderManageBloc extends RxBlocBase
       _mapToOnDeletedState();
 
   /// The state of [onUpdated] implemented in [_mapToOnUpdatedState]
-  late final ConnectableStream<Result<ReminderModel>> _onUpdatedState =
-      _mapToOnUpdatedState();
+  late final ConnectableStream<Result<IdentifiablePair<ReminderModel>>>
+      _onUpdatedState = _mapToOnUpdatedState();
 
   /// The state of [onCreated] implemented in [_mapToOnCreatedState]
   late final ConnectableStream<Result<ReminderModel>> _onCreatedState =
@@ -56,9 +56,13 @@ abstract class $ReminderManageBloc extends RxBlocBase
   void create(
           {required String title,
           required DateTime dueDate,
-          required bool complete}) =>
-      _$createEvent.add(
-          _CreateEventArgs(title: title, dueDate: dueDate, complete: complete));
+          required bool complete,
+          required bool completeUpdated}) =>
+      _$createEvent.add(_CreateEventArgs(
+          title: title,
+          dueDate: dueDate,
+          complete: complete,
+          completeUpdated: completeUpdated));
 
   @override
   void delete(ReminderModel reminder) => _$deleteEvent.add(reminder);
@@ -73,7 +77,8 @@ abstract class $ReminderManageBloc extends RxBlocBase
   ConnectableStream<Result<ReminderModel>> get onDeleted => _onDeletedState;
 
   @override
-  ConnectableStream<Result<ReminderModel>> get onUpdated => _onUpdatedState;
+  ConnectableStream<Result<IdentifiablePair<ReminderModel>>> get onUpdated =>
+      _onUpdatedState;
 
   @override
   ConnectableStream<Result<ReminderModel>> get onCreated => _onCreatedState;
@@ -84,7 +89,8 @@ abstract class $ReminderManageBloc extends RxBlocBase
 
   ConnectableStream<Result<ReminderModel>> _mapToOnDeletedState();
 
-  ConnectableStream<Result<ReminderModel>> _mapToOnUpdatedState();
+  ConnectableStream<Result<IdentifiablePair<ReminderModel>>>
+      _mapToOnUpdatedState();
 
   ConnectableStream<Result<ReminderModel>> _mapToOnCreatedState();
 
@@ -108,11 +114,16 @@ abstract class $ReminderManageBloc extends RxBlocBase
 /// [ReminderManageBlocEvents.create] event
 class _CreateEventArgs {
   const _CreateEventArgs(
-      {required this.title, required this.dueDate, required this.complete});
+      {required this.title,
+      required this.dueDate,
+      required this.complete,
+      required this.completeUpdated});
 
   final String title;
 
   final DateTime dueDate;
 
   final bool complete;
+
+  final bool completeUpdated;
 }