+++ packages/rx_bloc_list/lib/models.dart
@@ -2,5 +2,6 @@ library models;
 
 export 'src/extensions/identifiable_extensions.dart';
 export 'src/models/identifiable.dart';
+export 'src/models/identifiable_pair.dart';
 export 'src/models/managed_list.dart';
 export 'src/models/paginated_list.dart';