+++ packages/rx_bloc_list/lib/src/models/managed_list.dart
@@ -1,6 +1,6 @@
 import 'package:collection/collection.dart';
 
-import 'identifiable.dart';
+import '../../models.dart';
 
 /// The managed operation of a [Identifiable] object.
 /// - [merge] - Added or Updated
@@ -17,19 +17,20 @@ enum ManageOperation {
   ignore,
 }
 
-/// A class containing the list along with the managed [Identifiable] object.
+/// A class containing the list along with the managed [Identifiable] object:
+/// [updatedIdentifiable], which is part of [identifiablePair].
 class ManagedList<T extends Identifiable> {
   ManagedList(
     this.list, {
-    required this.identifiable,
+    required this.identifiablePair,
     required this.operation,
   });
 
-  /// The managed operation of the [identifiable] object
+  /// The managed operation of the [updatedIdentifiable] object
   final ManageOperation operation;
 
-  /// The managed object
-  final T identifiable;
+  /// The object containing the managed object [updatedIdentifiable]
+  final IdentifiablePair identifiablePair;
 
   /// The managed list
   final List<T> list;
@@ -39,13 +40,12 @@ class ManagedList<T extends Identifiable> {
     if (other is ManagedList<T>) {
       return const ListEquality().equals(other.list, list) &&
           operation == other.operation &&
-          identifiable.id == other.identifiable.id;
+          identifiablePair == other.identifiablePair;
     }
-
     return false;
   }
 
   @override
   int get hashCode =>
-      list.hashCode ^ operation.hashCode ^ identifiable.hashCode;
+      list.hashCode ^ operation.hashCode ^ identifiablePair.hashCode;
 }